refactor(api): extract schedule-time utils and sync openpath submodule
Verified-by: pre-commit
